Please create the PROBLEMS.md folder
This file is so you can get an idea of what bugs are there in the mod menu. And also some fixes for the bugs. If you create this, I think it would be a big help to you guys, knowing what bugs there are so you can fix them easily. In return,  I want nothing. So you have yourself a win win situation!