README changes
Update README.md

Upload fixed logo

Change README logo to Plain Logo

if we don't like it we can change it? but I like it.

Upload Simple Logo

This looks janky, this is only for notices.

Add status message file (see description)

Later I will make it so that the status message will appear on the cheat menu.